grader/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
redis==7.2.0
requests==2.32.3
websockets==16.0
pytest
pytest-asyncio